Move logic out of gateway
