+++ ProcessesApi.Tests/V1/Services/SoleToJointServiceTests.cs
@@ -16,6 +16,7 @@ using ProcessesApi.V1.Factories;
 using ProcessesApi.V1.Infrastructure.JWT;
 using Xunit;
 using ProcessesApi.V1.Services;
+using ProcessesApi.V1.Helpers;
 
 namespace ProcessesApi.Tests.V1.UseCase
 {
@@ -45,6 +46,7 @@ namespace ProcessesApi.Tests.V1.UseCase
         };
 
         private Mock<ISoleToJointGateway> _mockSTJGateway;
+        private Mock<ISoleToJointHelper> _mockHelper;
         private Mock<ISnsGateway> _mockSnsGateway;
 
         private readonly List<Action> _cleanup = new List<Action>();
@@ -75,7 +77,9 @@ namespace ProcessesApi.Tests.V1.UseCase
             _dbFixture = appFactory.DynamoDbFixture;
             _mockSTJGateway = new Mock<ISoleToJointGateway>();
             _mockSnsGateway = new Mock<ISnsGateway>();
-            _classUnderTest = new SoleToJointService(_mockSTJGateway.Object, new ProcessesSnsFactory(), _mockSnsGateway.Object);
+            _mockHelper = new Mock<ISoleToJointHelper>();
+
+            _classUnderTest = new SoleToJointService(_mockSTJGateway.Object, new ProcessesSnsFactory(), _mockSnsGateway.Object, _mockHelper.Object);
 
             _mockSnsGateway
                 .Setup(g => g.Publish(It.IsAny<EntityEventSns>(), It.IsAny<string>(), It.IsAny<string>()))
@@ -162,7 +166,7 @@ namespace ProcessesApi.Tests.V1.UseCase
         #region Automated eligibility checks
 
         [Fact]
-        public async Task AddIncomingTenantToRelatedEntitiesOnCheckEligibilityTrigger()
+        public async Task AddIncomingTenantToRelatedEntitiesOnCheckAutomatedEligibilityTrigger()
         {
             // Arrange
             var process = CreateProcessWithCurrentState(SoleToJointStates.SelectTenants);
@@ -176,6 +180,8 @@ namespace ProcessesApi.Tests.V1.UseCase
                                                         { SoleToJointFormDataKeys.IncomingTenantId, incomingTenantId },
                                                         { SoleToJointFormDataKeys.TenantId, tenantId },
                                                     });
+
+            _mockHelper.Setup(x => x.CheckAutomatedEligibility(process.TargetId, incomingTenantId, tenantId, _mockSTJGateway.Object)).ReturnsAsync(true);
             // Act
             await _classUnderTest.Process(triggerObject, process, _token).ConfigureAwait(false);
 
@@ -184,7 +190,7 @@ namespace ProcessesApi.Tests.V1.UseCase
         }
 
         [Fact]
-        public async Task CurrentStateIsUpdatedToAutomatedChecksFailedWhenCheckEligibilityReturnsFalse()
+        public async Task CurrentStateIsUpdatedToAutomatedChecksFailedWhenCheckAutomatedEligibilityReturnsFalse()
         {
             // Arrange
             var process = CreateProcessWithCurrentState(SoleToJointStates.SelectTenants);
@@ -200,7 +206,8 @@ namespace ProcessesApi.Tests.V1.UseCase
             var triggerObject = CreateProcessTrigger(process,
                                                      SoleToJointPermittedTriggers.CheckEligibility,
                                                      formData);
-            _mockSTJGateway.Setup(x => x.CheckEligibility(process.TargetId, incomingTenantId, tenantId)).ReturnsAsync(false);
+
+            _mockHelper.Setup(x => x.CheckAutomatedEligibility(process.TargetId, incomingTenantId, tenantId, _mockSTJGateway.Object)).ReturnsAsync(false);
             // Act
             await _classUnderTest.Process(triggerObject, process, _token).ConfigureAwait(false);
 
@@ -210,11 +217,11 @@ namespace ProcessesApi.Tests.V1.UseCase
                                                  SoleToJointStates.AutomatedChecksFailed,
                                                  new List<string>() { SoleToJointPermittedTriggers.CancelProcess });
             process.PreviousStates.LastOrDefault().State.Should().Be(SoleToJointStates.SelectTenants);
-            _mockSTJGateway.Verify(x => x.CheckEligibility(process.TargetId, incomingTenantId, tenantId), Times.Once());
+            _mockHelper.Verify(x => x.CheckAutomatedEligibility(process.TargetId, incomingTenantId, tenantId, _mockSTJGateway.Object), Times.Once());
         }
 
         [Fact]
-        public async Task ProcessStateIsUpdatedToAutomatedChecksPassedWhenCheckEligibilityReturnsTrue()
+        public async Task ProcessStateIsUpdatedToAutomatedChecksPassedWhenCheckAutomatedEligibilityReturnsTrue()
         {
             // Arrange
             var process = CreateProcessWithCurrentState(SoleToJointStates.SelectTenants);
@@ -230,7 +237,7 @@ namespace ProcessesApi.Tests.V1.UseCase
             var triggerObject = CreateProcessTrigger(process,
                                                      SoleToJointPermittedTriggers.CheckEligibility,
                                                      formData);
-            _mockSTJGateway.Setup(x => x.CheckEligibility(process.TargetId, incomingTenantId, tenantId)).ReturnsAsync(true);
+            _mockHelper.Setup(x => x.CheckAutomatedEligibility(process.TargetId, incomingTenantId, tenantId, _mockSTJGateway.Object)).ReturnsAsync(true);
 
             // Act
             await _classUnderTest.Process(triggerObject, process, _token).ConfigureAwait(false);
@@ -241,11 +248,11 @@ namespace ProcessesApi.Tests.V1.UseCase
                                                  SoleToJointStates.AutomatedChecksPassed,
                                                  new List<string>() { SoleToJointPermittedTriggers.CheckManualEligibility });
             process.PreviousStates.LastOrDefault().State.Should().Be(SoleToJointStates.SelectTenants);
-            _mockSTJGateway.Verify(x => x.CheckEligibility(process.TargetId, incomingTenantId, tenantId), Times.Once());
+            _mockHelper.Verify(x => x.CheckAutomatedEligibility(process.TargetId, incomingTenantId, tenantId, _mockSTJGateway.Object), Times.Once());
         }
 
         [Fact]
-        public void ThrowsFormDataNotFoundExceptionOnCheckEligibilityTrigger()
+        public void ThrowsFormDataNotFoundExceptionOnCheckAutomatedEligibilityTrigger()
         {
             // Arrange
             var process = CreateProcessWithCurrentState(SoleToJointStates.SelectTenants);
@@ -282,7 +289,7 @@ namespace ProcessesApi.Tests.V1.UseCase
 
             var snsEvent = new EntityEventSns();
 
-            _mockSTJGateway.Setup(x => x.CheckEligibility(process.TargetId, incomingTenantId, tenantId)).ReturnsAsync(false);
+            _mockHelper.Setup(x => x.CheckAutomatedEligibility(process.TargetId, incomingTenantId, tenantId, _mockSTJGateway.Object)).ReturnsAsync(false);
             _mockSnsGateway
                 .Setup(g => g.Publish(It.IsAny<EntityEventSns>(), It.IsAny<string>(), It.IsAny<string>()))
                 .Callback<EntityEventSns, string, string>((ev, s1, s2) => snsEvent = ev);
