+++ ProcessesApi/Startup.cs
@@ -40,6 +40,7 @@ using System.Linq;
 using System.Reflection;
 using System.Text.Json.Serialization;
 using ProcessesApi.V1;
+using ProcessesApi.V1.Helpers;
 
 namespace ProcessesApi
 {
@@ -156,6 +157,8 @@ namespace ProcessesApi
 
             RegisterGateways(services);
             RegisterUseCases(services);
+            RegisterHelpers(services);
+
             services.AddSingleton<IConfiguration>(Configuration);
 
             services.AddScoped<ISnsFactory, ProcessesSnsFactory>();
@@ -178,7 +181,6 @@ namespace ProcessesApi
             services.AddScoped<IProcessesGateway, ProcessesGateway>();
             services.AddScoped<ISoleToJointGateway, SoleToJointGateway>();
             services.AddScoped<IEntityUpdater, EntityUpdater>();
-
         }
 
         private static void RegisterUseCases(IServiceCollection services)
@@ -186,6 +188,11 @@ namespace ProcessesApi
             services.AddScoped<IGetByIdUseCase, GetProcessByIdUseCase>();
             services.AddScoped<IProcessUseCase, ProcessUseCase>();
             services.AddScoped<IUpdateProcessByIdUsecase, UpdateProcessByIdUsecase>();
+        }
+
+        private static void RegisterHelpers(IServiceCollection services)
+        {
+            services.AddScoped<ISoleToJointHelper, SoleToJointHelper>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
