--- ProcessesApi/V1/Helpers/SoleToJointHelpers.cs
@@ -1,7 +1,0 @@
-namespace ProcessesApi.V1.Helpers
-{
-    public class SoleToJointHelpers
-    {
-
-    }
-}
